chore: refactor and clean up by moving stuff out of App
