+++ src/components/app/data/hooks/useUserEntitlements.js
@@ -1,0 +1,10 @@
+import { useQuery } from '@tanstack/react-query';
+import { makeUserEntitlementsQuery } from '../../routes/queries';
+
+/**
+ * Retrieves the user entitlements.
+ * @returns {Types.UseQueryResult}} The query results for the user entitlements.
+ */
+export default function useUserEntitlements() {
+  return useQuery(makeUserEntitlementsQuery());
+}
