+++ src/components/app/routes/loaders/index.js
@@ -1,0 +1,4 @@
+export { default as makeUpdateActiveEnterpriseCustomerUserLoader } from './updateActiveEnterpriseCustomerUserLoader';
+export { default as makeRootLoader } from './rootLoader';
+export { default as makeCourseLoader } from './courseLoader';
+export { default as makeDashboardLoader } from './dashboardLoader';
