+++ src/components/site-header/SiteHeader.jsx
@@ -7,8 +7,8 @@ import { Container, MediaQuery } from '@edx/paragon';
 import SiteHeaderLogos from './SiteHeaderLogos';
 import SiteHeaderNavMenu from './SiteHeaderNavMenu';
 import { Menu, MenuTrigger, MenuContent } from './menu';
+import { useEnterpriseLearner } from '../app/data';
 import AvatarDropdown from './AvatarDropdown';
-import { useEnterpriseLearner } from '../app/App';
 
 const SiteHeader = () => {
   const config = getConfig();
