+++ src/components/site-header/SiteHeaderNavMenu.jsx
@@ -1,7 +1,6 @@
 import { useIntl } from '@edx/frontend-platform/i18n';
 import { NavLink } from 'react-router-dom';
-
-import { useEnterpriseLearner } from '../app/App';
+import { useEnterpriseLearner } from '../app/data';
 
 const SiteHeaderNavMenu = () => {
   const { data: { enterpriseCustomer } } = useEnterpriseLearner();
