+++ src/types.d.ts
@@ -1,0 +1,7 @@
+/**
+ * @typedef {import('@tanstack/react-query').UseQueryResult} UseQueryResult
+ */
+
+export type UseQueryResult = import('@tanstack/react-query').UseQueryResult;
+
+export as namespace Types;
